requirements.txt
openai
python-dotenv
groq
flask
